art: Copy simple_click_cursor into art
Make a bigger version of simple_click_cursor so we can use high res art
and offset fingers. Give it a simpler name since it's our only cursor.

Need to copy and setup all the anims again because the slicing is
different and godot was remembering the old assets.

Changed the active, point, and inactive cursors.

Delete unused cursor_active. I'm animating it instead.