Popochiu: Layout dialogue above heads and wrapped
This wasn't constraining the width of the text, so set x as well. Use
the target to figure out where to position, but we need to adjust for
the size of the text so it still lines up with the "Dialog Position".
There's some default height and using 150 makes it line up okay.